增加表单管理的筛选条件

+++ api/src/main/java/com/qu/modules/web/vo/QuestionAndCategoryPageVo.java
@@ -1,16 +1,16 @@
 package com.qu.modules.web.vo;
 
+import java.util.List;
+
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
-
-import java.util.List;
 
 @Data
 @ApiModel(value = "问卷分页Vo", description = "问卷分页Vo")
 public class QuestionAndCategoryPageVo {
     @ApiModelProperty(value = "总条数")
-    private int total;
+    private long total;
     @ApiModelProperty(value = "数据")
     private List<QuestionAndCategoryVo> questionList;
 }
